requirements.txt
AutoRAG[ko,ja,gpu]>=0.3.5
pyyaml